src/ai_foundry/requirements.txt
azure-identity
azure-ai-agents
azure-ai-projects
python-dotenv